+++ intel_extension_for_transformers/transformers/dpo_trainer.py
@@ -29,7 +29,7 @@ import importlib
 from peft import PeftModel, get_peft_model, prepare_model_for_kbit_training
 import time
 import logging
-from intel_extension_for_transformers.neural_chat.utils.common import is_hpu_available
+from intel_extension_for_transformers.tools.utils import is_hpu_available
 
 logger = logging.getLogger(__name__)
