+++ intel_extension_for_transformers/transformers/llm/quantization/utils.py
@@ -26,7 +26,7 @@ from neural_compressor import quantization
 from neural_compressor.adaptor.torch_utils.model_wrapper import WeightOnlyLinear
 from neural_compressor.utils.utility import LazyImport
 from neural_compressor.config import PostTrainingQuantConfig
-from intel_extension_for_transformers.neural_chat.utils.common import is_ipex_available, is_autoround_available
+from intel_extension_for_transformers.tools.utils import is_ipex_available, is_autoround_available
 from transformers import AutoTokenizer
 
 if is_ipex_available():
