+++ requirements.txt
@@ -1,6 +1,3 @@
-fastapi
 py-cpuinfo
 setuptools>=65
 setuptools_scm[toml]>=6.2
-uvicorn
-yacs
