+++ tests/CI/test_weight_only_gpu.py
@@ -24,7 +24,7 @@ from intel_extension_for_transformers.transformers.modeling import AutoModelForC
 from intel_extension_for_transformers.transformers import GPTQConfig, RtnConfig
 from math import isclose
 from transformers import AutoTokenizer
-from intel_extension_for_transformers.neural_chat.utils.common import get_gpu_family, is_ipex_available
+from intel_extension_for_transformers.tools.utils import get_gpu_family, is_ipex_available
 from torch.utils.data import DataLoader
